requirements.txt
PyQt6>=6.7
pyserial>=3.5
